chore: stop exposing internal actions/state/reducer to the builder
