+++ .changeset/grumpy-poets-compare.md
@@ -1,0 +1,5 @@
+---
+"@makeswift/runtime": patch
+---
+
+chore: stop exposing internal actions/state/reducer to the builder
