+++ packages/runtime/src/builder/breakpoints/index.ts
@@ -1,8 +1,6 @@
 export {
-  reducer,
   DefaultBreakpointID,
   DEFAULT_BREAKPOINTS,
-  type State,
   type BreakpointId,
   type Breakpoint,
   type Breakpoints,
