+++ packages/runtime/src/state/modules/locale.ts
@@ -1,5 +1,5 @@
 import { type Action, type UnknownAction, isKnownAction } from '../actions'
-import { SharedActionTypes } from '../shared-api'
+import { ReadOnlyActionTypes } from '../actions/internal/read-only-actions'
 
 export type State = string | null
 
@@ -15,7 +15,7 @@ export function reducer(state = getInitialState(), action: Action | UnknownActio
   if (!isKnownAction(action)) return state
 
   switch (action.type) {
-    case SharedActionTypes.SET_LOCALE:
+    case ReadOnlyActionTypes.SET_LOCALE:
       return action.payload.locale
 
     default:
